Strict item name matching
This will allow us to handle very malformed bids on
items (no spaces between item and bid).
I'm not entirely sure if it's worth the trouble since we have to
introduce some new tracking logic into the gui thread that's
orchestrating the process. I made some structural changes so we
aren't matching the same line twice.

You could do an alternate approach where the parser becomes an
object and gains state, so it could track it's own list of opened auctions
but that seems to muddy the logic even more.